Move some Windows-related state tests to pytest
